Move selectedCurrency to simple redux state Remove currencyValues redux state
