+++ src/logic/safe/store/actions/setSelectedCurrency.ts
@@ -1,0 +1,5 @@
+import { createAction } from 'redux-actions'
+
+export const SET_CURRENT_CURRENCY = 'SET_CURRENT_CURRENCY'
+
+export const setSelectedCurrency = createAction(SET_CURRENT_CURRENCY)
