+++ src/routes/safe/components/Balances/Coins/index.tsx
@@ -14,7 +14,6 @@ import Table from 'src/components/Table'
 import { cellWidth } from 'src/components/Table/TableHead'
 import Button from 'src/components/layout/Button'
 import Row from 'src/components/layout/Row'
-import { currentCurrencySelector } from 'src/logic/currencyValues/store/selectors'
 import { BALANCE_ROW_TEST_ID } from 'src/routes/safe/components/Balances'
 import AssetTableCell from 'src/routes/safe/components/Balances/AssetTableCell'
 import {
@@ -29,6 +28,7 @@ import { extendedSafeTokensSelector, grantedSelector } from 'src/routes/safe/con
 import { useAnalytics, SAFE_NAVIGATION_EVENT } from 'src/utils/googleAnalytics'
 import { makeStyles } from '@material-ui/core/styles'
 import { styles } from './styles'
+import { currentCurrencySelector } from 'src/logic/safe/store/selectors'
 
 const useStyles = makeStyles(styles)
